Rename create/restore account fragment classes
