+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/manageaccounts/ManageAccountsFragment.kt
@@ -121,12 +121,12 @@ fun ManageAccountsScreen(navController: NavController, mode: ManageAccountsModul
                     val actions = listOf(
                         ActionViewItem(R.drawable.ic_plus, R.string.ManageAccounts_CreateNewWallet) {
                             navController.navigateWithTermsAccepted {
-                                navController.slideFromRight(R.id.createAccountIntroFragment, args)
+                                navController.slideFromRight(R.id.createAccountFragment, args)
                             }
                         },
                         ActionViewItem(R.drawable.ic_download_20, R.string.ManageAccounts_ImportWallet) {
                             navController.navigateWithTermsAccepted {
-                                navController.slideFromRight(R.id.restoreFragment, args)
+                                navController.slideFromRight(R.id.restoreAccountFragment, args)
                             }
                         },
                         ActionViewItem(R.drawable.icon_binocule_20, R.string.ManageAccounts_WatchAddress) {
